working with state as Double, saving with graph

+++ src/main/scala/Command.scala
@@ -1,10 +1,10 @@
 object Command {
   sealed trait Command
 
-  case class Add(amount: Int) extends Command
+  case class Add(amount: Double) extends Command
 
-  case class Multiply(amount: Int) extends Command
+  case class Multiply(amount: Double) extends Command
 
-  case class Divide(amount: Int) extends Command
+  case class Divide(amount: Double) extends Command
 
 }
